Fixed UT compilation. (#60)
